+++ com-mmxlabs-models-lng/com.mmxlabs.models.lng.schedule/src/com/mmxlabs/models/lng/schedule/FuelQuantity.java
@@ -63,12 +63,12 @@ public interface FuelQuantity extends EObject {
 	 * </p>
 	 * <!-- end-user-doc -->
 	 * @return the value of the '<em>Cost</em>' attribute.
-	 * @see #setCost(BigDecimal)
+	 * @see #setCost(int)
 	 * @see com.mmxlabs.models.lng.schedule.SchedulePackage#getFuelQuantity_Cost()
 	 * @model required="true"
 	 * @generated
 	 */
-	BigDecimal getCost();
+	int getCost();
 
 	/**
 	 * Sets the value of the '{@link com.mmxlabs.models.lng.schedule.FuelQuantity#getCost <em>Cost</em>}' attribute.
@@ -78,7 +78,7 @@ public interface FuelQuantity extends EObject {
 	 * @see #getCost()
 	 * @generated
 	 */
-	void setCost(BigDecimal value);
+	void setCost(int value);
 
 	/**
 	 * Returns the value of the '<em><b>Amounts</b></em>' containment reference list.
